requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
evaluate==0.4.3
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.55.4
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
GPUtil==1.4.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.2
nvidia-pyindex==1.0.9
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
peft==0.14.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
stevedore==5.4.0
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wmi==1.5.1
xxhash==3.5.0
yarl==1.18.3
speedtest-cli==2.1.3
mss>=9.0.1  # Screen capture

# Voice and Speech Dependencies
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13

# Voice Service Dependencies
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.13
sounddevice>=0.4.6
vosk>=0.3.45

# Vision Dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
scikit-learn>=1.3.0
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
scipy>=1.11.0
matplotlib>=3.7.0

# Vision Service Dependencies
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
mediapipe>=0.10.0
pillow>=10.0.0
face-recognition>=1.3.0
dlib>=19.24.0

# Code Analysis Dependencies
autopep8>=2.0.0
pylint>=3.0.0
docstring-parser>=0.15
astroid>=3.0.0

# Media Processing Dependencies
pydub>=0.25.1
moviepy>=1.0.3
mutagen>=1.46.0

# System Monitoring Dependencies
psutil>=5.9.0
wmi>=1.5.0
py-cpuinfo>=9.0.0

# File Operations Dependencies
send2trash>=1.8.0
pathspec>=0.11.0
watchdog>=3.0.0

# RSS feed parser dependencies
feedparser==6.0.10
beautifulsoup4==4.12.3

# Refactored Structure Dependencies
dependency-injector>=4.41.0
pillow>=10.0.0
pyyaml>=6.0.1
dataclasses-json>=0.5.7

# Memory Service Dependencies
rapidfuzz>=3.0.0
sentence-transformers>=2.2.2
zstandard>=0.21.0
numpy>=1.24.0
pandas>=2.0.0

# Advanced Vision Models
transformers>=4.30.0  # For Hugging Face models
timm>=0.9.0  # PyTorch Image Models
midas-py>=1.0.0  # MiDaS depth estimation
segment-anything>=1.0  # Meta's Segment Anything Model
supervision>=0.15.0  # Vision annotations and tracking
onnxruntime>=1.15.0  # ONNX Runtime for optimized inference
